# backend/swagger.lock.json
{"openapi":"3.0.0","paths":{"/v1/ventas":{"post":{"operationId":"VentasController_crear","parameters":[],"responses":{"201":{"description":""}}},"get":{"operationId":"VentasController_listar","parameters":[{"name":"estado","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/v1/ventas/{codigo}":{"get":{"operationId":"VentasController_detalle","parameters":[{"name":"codigo","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/v1/ventas/{codigo}/evidencias":{"post":{"operationId":"VentasController_evidencias","parameters":[{"name":"codigo","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/v1/ventas/{codigo}/asegurar-ins":{"post":{"operationId":"VentasController_asegurarIns","parameters":[{"name":"codigo","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/v1/ventas/{codigo}/pagar":{"post":{"operationId":"VentasController_pagar","parameters":[{"name":"codigo","required":true,"in":"path","schema":{"type":"string"}},{"name":"idempotency-key","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/health":{"get":{"operationId":"HealthController_health","parameters":[],"responses":{"200":{"description":""}}}},"/v1/catalogos/municipios":{"get":{"operationId":"CatalogosController_municipios","parameters":[{"name":"q","required":true,"in":"query","schema":{"type":"string"}},{"name":"activos","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/v1/catalogos/vias":{"get":{"operationId":"CatalogosController_vias","parameters":[{"name":"q","required":true,"in":"query","schema":{"type":"string"}},{"name":"activos","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/v1/catalogos/sectores":{"get":{"operationId":"CatalogosController_sectores","parameters":[{"name":"municipio","required":true,"in":"query","schema":{"type":"string"}},{"name":"zona","required":true,"in":"query","schema":{"type":"string"}},{"name":"q","required":true,"in":"query","schema":{"type":"string"}},{"name":"activos","required":true,"in":"query","schema":{"type":"string"}},{"name":"format","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/v1/catalogos/motivos-reagenda":{"get":{"operationId":"MotivosReagendaPublicControllerKebab_listarKebab","parameters":[],"responses":{"200":{"description":""}}}},"/v1/catalogos/motivos_reagenda/items":{"get":{"operationId":"MotivosReagendaPublicControllerUnderscore_listarUnderscore","parameters":[],"responses":{"200":{"description":""}}}},"/v1/admin/catalogos/motivos_reagenda/items":{"get":{"operationId":"MotivosReagendaAdminController_listarAdmin","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"MotivosReagendaAdminController_crear","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCatalogoItemDto"}}}},"responses":{"201":{"description":""}}}},"/v1/admin/catalogos/motivos_reagenda/items/{id}":{"patch":{"operationId":"MotivosReagendaAdminController_actualizar","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCatalogoItemDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"MotivosReagendaAdminController_eliminar","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/v1/catalogos/motivos-anulacion":{"get":{"operationId":"MotivosAnulacionPublicControllerKebab_list","parameters":[],"responses":{"200":{"description":""}}}},"/v1/admin/catalogos/motivos_anulacion/items":{"get":{"operationId":"MotivosAnulacionAdminController_listAll","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"MotivosAnulacionAdminController_create","parameters":[],"responses":{"201":{"description":""}}}},"/v1/admin/catalogos/motivos_anulacion/items/{id}":{"patch":{"operationId":"MotivosAnulacionAdminController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"delete":{"operationId":"MotivosAnulacionAdminController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/v1/agenda/ordenes/{codigo}/asignar":{"post":{"operationId":"AgendaController_asignar","parameters":[{"name":"codigo","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/v1/agenda/ordenes/{codigo}/reagendar":{"post":{"operationId":"AgendaController_reagendar","parameters":[{"name":"codigo","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/v1/agenda/ordenes/{codigo}/cancelar":{"post":{"operationId":"AgendaController_cancelar","parameters":[{"name":"codigo","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/v1/agenda/ordenes/{codigo}/anular":{"post":{"operationId":"AgendaController_anular","parameters":[{"name":"codigo","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/v1/tecnicos/{tecnicoId}/ordenes/pendientes":{"get":{"operationId":"TecnicosController_pendientes","summary":"Órdenes pendientes (agendada/en_progreso) para un técnico","parameters":[{"name":"tecnicoId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["tecnicos"]}},"/v1/tecnicos/{tecnicoId}/ordenes/{ordenId}/iniciar":{"post":{"operationId":"TecnicosController_iniciarPorId","summary":"Inicia una orden por ID (UUID) — idempotente","parameters":[{"name":"tecnicoId","required":true,"in":"path","schema":{"type":"string"}},{"name":"ordenId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["tecnicos"]}},"/v1/tecnicos/{tecnicoId}/ordenes/codigo/{codigo}/iniciar":{"post":{"operationId":"TecnicosController_iniciarPorCodigo","summary":"Inicia una orden por código — idempotente","parameters":[{"name":"tecnicoId","required":true,"in":"path","schema":{"type":"string"}},{"name":"codigo","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["tecnicos"]}},"/v1/tecnicos/{tecnicoId}/ordenes/{ordenId}/cerrar":{"post":{"operationId":"TecnicosController_cerrarPorId","summary":"Cierra una orden por ID (UUID), aplica materiales, firma y evidencias — idempotente","parameters":[{"name":"tecnicoId","required":true,"in":"path","schema":{"type":"string"}},{"name":"ordenId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CerrarOrdenDto"}}}},"responses":{"201":{"description":""}},"tags":["tecnicos"]}},"/v1/tecnicos/{tecnicoId}/ordenes/codigo/{codigo}/cerrar":{"post":{"operationId":"TecnicosController_cerrarPorCodigo","summary":"Cierra una orden por código, aplica materiales, firma y evidencias — idempotente","parameters":[{"name":"tecnicoId","required":true,"in":"path","schema":{"type":"string"}},{"name":"codigo","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CerrarOrdenDto"}}}},"responses":{"201":{"description":""}},"tags":["tecnicos"]}},"/v1/pdf/probe":{"get":{"operationId":"PdfController_probeGet","parameters":[],"responses":{"200":{"description":""}}}},"/v1/pdf/probe-put":{"post":{"operationId":"PdfController_probePut","parameters":[{"name":"key","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/v1/inventario/stock/almacen/{codigo}":{"get":{"operationId":"InventarioController_stockPorAlmacen","summary":"Stock por almacén (código)","parameters":[{"name":"codigo","required":true,"in":"path","schema":{"type":"string"}},{"name":"materialId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Inventario"]}},"/v1/inventario/tecnicos/{id}/stock":{"get":{"operationId":"InventarioController_stockTecnico","summary":"Stock de almacén técnico TEC-{id}","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Inventario"]}},"/v1/inventario/kardex":{"get":{"operationId":"InventarioController_kardex","summary":"Últimos movimientos (kardex)","parameters":[{"name":"limit","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Inventario"]}},"/v1/inventario/movimientos":{"post":{"operationId":"InventarioController_crearMovimiento","summary":"Crear movimiento (ingreso/egreso/traslado/ajuste)","description":"Para traslado puedes enviar tipo='traslado' o 'transferencia' (se normaliza a 'traslado').","parameters":[{"name":"idempotency-key","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Inventario"]}},"/v1/inventario/tecnicos/{id}/agregar":{"post":{"operationId":"InventarioController_agregarTecnico","summary":"Trasladar desde MAIN a TEC-{id}","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"idempotency-key","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Inventario"]}},"/v1/inventario/tecnicos/{id}/descontar":{"post":{"operationId":"InventarioController_descontarTecnico","summary":"Devolver desde TEC-{id} a MAIN","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"idempotency-key","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Inventario"]}},"/v1/inventario/tecnicos/{id}/ajustar":{"post":{"operationId":"InventarioController_ajustarTecnico","summary":"Ajuste sobre TEC-{id}","description":"signo='mas' suma stock, signo='menos' descuenta stock","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"idempotency-key","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Inventario"]}},"/v1/inventario/kardex/material":{"get":{"operationId":"KardexMaterialController_alias","summary":"Alias 307 → /v1/inventario/kardex con filtros por material","parameters":[{"name":"materialId","required":true,"in":"query","description":"ID numérico del material","schema":{"type":"string"}},{"name":"from","required":false,"in":"query","description":"ISO date desde (incl.)","schema":{"type":"string"}},{"name":"to","required":false,"in":"query","description":"ISO date hasta (incl.)","schema":{"type":"string"}},{"name":"almacen","required":false,"in":"query","description":"Código de almacén (ej: CENTRAL, ALM-PRINC, TEC-6)","schema":{"type":"string"}},{"name":"tecnicoId","required":false,"in":"query","description":"Filtrar por técnico (numérico)","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"Límite de filas (1..500). Default 50","schema":{"type":"string"}}],"responses":{"307":{"description":"Redirect temporal al kardex canónico"}},"tags":["Inventario"]}},"/v1/materiales":{"get":{"operationId":"MaterialesController_list","parameters":[],"responses":{"200":{"description":""}}}},"/v1/materiales/disponibles":{"get":{"operationId":"MaterialesDisponiblesController_disponibles","summary":"Alias a inventario/stock por almacén (solo lectura)","description":"Retorna los materiales disponibles en un almacén. Internamente redirige (307) a /v1/inventario/stock/almacen/:codigo.","parameters":[{"name":"almacen","required":true,"in":"query","schema":{"example":"CENTRAL","type":"string"}}],"responses":{"307":{"description":"Redirect to /v1/inventario/stock/almacen/:codigo"},"400":{"description":"Parámetros inválidos"}},"tags":["Materiales"]}},"/v1/ordenes":{"get":{"operationId":"OrdenesController_list","parameters":[],"responses":{"200":{"description":""}}}},"/v1/ordenes/{codigo}":{"get":{"operationId":"OrdenesController_byCodigo","parameters":[{"name":"codigo","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/v1/ordenes/{codigo}/guardar":{"put":{"operationId":"OrdenesController_guardar","parameters":[{"name":"codigo","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/v1/ordenes/{codigo}/evidencias":{"post":{"operationId":"OrdenesController_evidencias","parameters":[{"name":"codigo","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/v1/ordenes/{codigo}/cerrar":{"put":{"operationId":"OrdenesController_cerrar","parameters":[{"name":"codigo","required":true,"in":"path","schema":{"type":"string"}},{"name":"Idempotency-Key","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"post":{"operationId":"OrdenesController_cerrarPost","parameters":[{"name":"codigo","required":true,"in":"path","schema":{"type":"string"}},{"name":"Idempotency-Key","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/v1/metrics":{"get":{"operationId":"MetricsController_getMetrics","parameters":[],"responses":{"200":{"description":""}}}},"/v1/jobs/simular-cortes":{"post":{"operationId":"JobsController_simularCortes","summary":"Simular cortes (safe/dry-run disponible)","parameters":[],"responses":{"202":{"description":""}},"tags":["Jobs"]}},"/v1/jobs/simular-reconexiones":{"post":{"operationId":"JobsController_simularReconexiones","summary":"Simular reconexiones (safe/dry-run disponible)","parameters":[],"responses":{"202":{"description":""}},"tags":["Jobs"]}},"/v1/equipos/disponibles":{"get":{"operationId":"EquiposController_disponibles","summary":"Listar equipos disponibles por tipo","parameters":[{"name":"tipo","required":true,"in":"query","schema":{"enum":["ONU","REPETIDOR","ONT"],"type":"string"}},{"name":"tecnico","required":false,"in":"query","description":"Reservado; hoy se ignora","schema":{}}],"responses":{"200":{"description":"OK"},"400":{"description":"Parámetros inválidos"}},"tags":["Equipos"]}},"/v1/equipos/entregar":{"post":{"operationId":"EquiposController_entregar","summary":"Entregar equipo a técnico (ALMACEN -> TECNICO), idempotente","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["id","tecnicoId"],"properties":{"id":{"type":"string","format":"uuid"},"tecnicoId":{"type":"integer"},"fromAlmacen":{"type":"string","example":"ALM-PRINC"}}}}}},"responses":{"201":{"description":""}},"tags":["Equipos"]}},"/v1/equipos/devolver":{"post":{"operationId":"EquiposController_devolver","summary":"Devolver equipo a almacén (TECNICO -> ALMACEN), idempotente","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["id"],"properties":{"id":{"type":"string","format":"uuid"},"tecnicoId":{"type":"integer","nullable":true},"toAlmacen":{"type":"string","example":"ALM-PRINC"}}}}}},"responses":{"201":{"description":""}},"tags":["Equipos"]}},"/v1/equipos/stock":{"get":{"operationId":"EquiposController_stock","summary":"Listar stock de equipos (por almacén o por técnico)","parameters":[{"name":"almacen","required":false,"in":"query","description":"Código de almacén (p.ej. ALM-PRINC)","schema":{"type":"string"}},{"name":"tecnicoId","required":false,"in":"query","description":"Id de técnico (numérico)","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Equipos"]}},"/v1/equipos/historial":{"get":{"operationId":"EquiposHistorialController_historial","summary":"Consultar equipo por SN o MAC (read-only)","description":"Devuelve el equipo actual por sn o mac. El arreglo \"historial\" queda vacío por ahora (sin tabla de movimientos).","parameters":[{"name":"sn","required":false,"in":"query","schema":{"type":"string"}},{"name":"mac","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"OK"},"400":{"description":"Parámetros inválidos"},"404":{"description":"No encontrado"}},"tags":["Equipos"]}},"/v1/equipos/reservar":{"post":{"operationId":"EquiposReservasController_reservar","parameters":[],"responses":{"201":{"description":""}}}},"/v1/equipos/liberar":{"post":{"operationId":"EquiposReservasController_liberar","parameters":[],"responses":{"201":{"description":""}}}},"/v1/equipos/reservas":{"get":{"operationId":"EquiposReservasController_reservas","parameters":[{"name":"tecnicoId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"ISP Backend API","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"x-api-key":{"type":"apiKey","in":"header","name":"x-api-key"}},"schemas":{"CreateCatalogoItemDto":{"type":"object","properties":{}},"UpdateCatalogoItemDto":{"type":"object","properties":{}},"CerrarOrdenDto":{"type":"object","properties":{}}}}}